move the Policy builder to its own file
